requirements.txt
aiogram==3.22.0
python-dotenv==1.1.1
sqlalchemy==2.0.43
asyncpg==0.30.0